Revert "Merge branch 'development' into feature/Advait/CalendarCard"
This reverts commit 7438804b124476d7397a442639829c5d4a9d9f70, reversing
changes made to 24ead10203bb51a93600429ed45ea1a9970bedc9.
