+++ package.json
@@ -15,8 +15,7 @@
     "react": "^18.3.1",
     "react-datepicker": "^8.3.0",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^6.22.3",
-    "react-select": "^5.10.1"
+    "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.1",
