+++ src/components/layout/RoleLayout.tsx
@@ -10,8 +10,7 @@ import {
   Settings, 
   LogOut, 
   ChevronLeft,
-  User,
-  PlusCircle
+  User
 } from 'lucide-react';
 
 interface NavItem {
@@ -26,20 +25,19 @@ const navConfig: Record<string, NavItem[]> = {
     { label: 'Travel Requests', path: '/admin/travel-requests', icon: <Briefcase className="h-5 w-5" /> },
     { label: 'Calendar', path: '/admin/calendar', icon: <Calendar className="h-5 w-5" /> },
     { label: 'Reports', path: '/admin/reports', icon: <BarChart className="h-5 w-5" /> },
-    // { label: 'Settings', path: '/admin/settings', icon: <Settings className="h-5 w-5" /> }
+    { label: 'Settings', path: '/admin/settings', icon: <Settings className="h-5 w-5" /> }
   ],
   manager: [
     { label: 'Dashboard', path: '/manager/dashboard', icon: <Home className="h-5 w-5" /> },
     { label: 'Team Requests', path: '/manager/team-requests', icon: <Briefcase className="h-5 w-5" /> },
     { label: 'My Requests', path: '/manager/my-requests', icon: <Briefcase className="h-5 w-5" /> },
-    // { label: 'Calendar', path: '/manager/calendar', icon: <Calendar className="h-5 w-5" /> }
+    { label: 'Calendar', path: '/manager/calendar', icon: <Calendar className="h-5 w-5" /> }
   ],
   employee: [
     { label: 'Dashboard', path: '/employee/dashboard', icon: <Home className="h-5 w-5" /> },
-    // { label: 'New Request', path: '/employee/new-request', icon: <Briefcase className="h-5 w-5" /> },
+    { label: 'New Request', path: '/employee/new-request', icon: <Briefcase className="h-5 w-5" /> },
     { label: 'My Requests', path: '/employee/my-requests', icon: <Briefcase className="h-5 w-5" /> },
-    // { label: 'New Request', path: '/employee/create-request', icon: <PlusCircle className="h-5 w-5" /> },
-    // { label: 'Calendar', path: '/employee/calendar', icon: <Calendar className="h-5 w-5" /> }
+    { label: 'Calendar', path: '/employee/calendar', icon: <Calendar className="h-5 w-5" /> }
   ]
 };
